refactor to move router switch to app. trimmed nav component by close to 200 lines. Much cleaner
