+++ client/src/App.js
@@ -1,12 +1,30 @@
 import React, { useEffect, useState } from 'react';
+import {
+  pageVariants,
+  pageTransition,
+  pageStyle,
+} from './utils/animationTransitions';
+import { Switch, Route, useLocation, Redirect } from 'react-router-dom';
+import { AnimatePresence } from 'framer-motion';
 import './App.css';
-import Footer from './components/footer';
 import Header from './components/header';
+import {
+  DigitalDesign,
+  PrintDesign,
+  GraphicDesign,
+  UIUX,
+  Information,
+} from './components/pages';
 import StickySocials from './components/sticky_socials';
+import Footer from './components/footer';
 
 const App = () => {
   const [isDark, setDark] = useState(false);
   const [isSticky, setSticky] = useState(false);
+  const [variant, setVariant] = useState(pageVariants);
+
+  const location = useLocation();
+
   useEffect(() => {
     const nav = document.querySelector('nav');
     const stickyLogo = document.querySelector('.sticky-logo-cont');
@@ -51,8 +69,64 @@ const App = () => {
 
   return (
     <div className='App'>
-      <Header updateMode={updateMode} />
-      {/* header holds nav component which renders each page and holds all routing logic for transitions */}
+      <Header
+        updateMode={updateMode}
+        variant={variant}
+        setVariant={setVariant}
+        location={location}
+      />
+      <AnimatePresence>
+        <Switch location={location} key={location.pathname}>
+          <Route exact path='/'>
+            <Redirect to='/digital_design' />
+          </Route>
+          <Route path='/digital_design'>
+            <div className='page-wrap'>
+              <DigitalDesign
+                pageVariants={variant}
+                pageStyle={pageStyle}
+                pageTransition={pageTransition}
+              />
+            </div>
+          </Route>
+          <Route path='/print_design'>
+            <div className='page-wrap'>
+              <PrintDesign
+                pageVariants={variant}
+                pageStyle={pageStyle}
+                pageTransition={pageTransition}
+              />
+            </div>
+          </Route>
+          <Route path='/graphic_design'>
+            <div className='page-wrap'>
+              <GraphicDesign
+                pageVariants={variant}
+                pageStyle={pageStyle}
+                pageTransition={pageTransition}
+              />
+            </div>
+          </Route>
+          <Route path='/ui_ux'>
+            <div className='page-wrap'>
+              <UIUX
+                pageVariants={variant}
+                pageStyle={pageStyle}
+                pageTransition={pageTransition}
+              />
+            </div>
+          </Route>
+          <Route path='/information'>
+            <div className='page-wrap'>
+              <Information
+                pageVariants={variant}
+                pageStyle={pageStyle}
+                pageTransition={pageTransition}
+              />
+            </div>
+          </Route>
+        </Switch>
+      </AnimatePresence>
       {isSticky && <StickySocials updateMode={updateMode} />}
       {isSticky && <Footer />}
     </div>
