+++ client/src/utils/animationTransitions.js
@@ -1,0 +1,45 @@
+export const pageVariants = {
+  initial: {
+    opacity: 0,
+    x: '200vw',
+    scale: 1,
+  },
+  in: {
+    opacity: 1,
+    x: 0,
+    scale: 1,
+  },
+  out: {
+    opacity: 0,
+    // x: '-100vw',
+    scale: 1,
+  },
+};
+
+export const pageVariantsAlt = {
+  initial: {
+    opacity: 0,
+    x: '-100vw',
+    scale: 1,
+  },
+  in: {
+    opacity: 1,
+    x: 0,
+    scale: 1,
+  },
+  out: {
+    opacity: 0,
+    // x: '200vw',
+    scale: 1,
+  },
+};
+
+export const pageTransition = {
+  type: 'tween',
+  ease: 'easeInOut',
+  duration: 0.7,
+};
+
+export const pageStyle = {
+  position: 'absolute',
+};
